samostatný soubor pro medkit
NEFUNKČNÍ
