+++ source/client/main.py
@@ -7,6 +7,7 @@ import time
 import sys
 import os
 import math
+from medkit import Medkit
 
 # Pro server hostovaný na Render.com použij:
 SERVER_URL = "wss://projekt-1ep-tabor.onrender.com/ws"
@@ -424,33 +425,6 @@ def shoot(weapon_name):
     print(f"Střelba ze zbraně: {weapon_name}, poškození: {WEAPONS[weapon_name]['damage']}")
     
     return True
-
-class Medkit:
-    def __init__(self, x=0, y=0, size=80):
-        self.x = x  # Pozice na mapě (ne na obrazovce)
-        self.y = y  # Pozice na mapě (ne na obrazovce)
-        self.size = size
-
-        try:
-            self.image = pygame.image.load("images/medkit.png").convert_alpha()
-            self.image = pygame.transform.scale(self.image, (size, size))
-        except Exception as e:
-            print(f"Chyba při načítání medkitu: {e}")
-            self.image = pygame.Surface((size, size), pygame.SRCALPHA)
-            pygame.draw.rect(self.image, (255, 0, 0), (0, 0, size, size))
-            pygame.draw.rect(self.image, (255, 255, 255), (size//4, size//4, size//2, size//2))
-            pygame.draw.line(self.image, (255, 255, 255), (size//2, size//4), (size//2, size*3//4), 3)
-    
-    def draw(self, screen, camera_x, camera_y):
-        screen_x = int(self.x - camera_x + SCREEN_WIDTH // 2)
-        screen_y = int(self.y - camera_y + SCREEN_HEIGHT // 2)
-        
-        if (-self.size <= screen_x <= SCREEN_WIDTH + self.size and 
-            -self.size <= screen_y <= SCREEN_HEIGHT + self.size):
-            screen.blit(self.image, (screen_x - self.size // 2, screen_y - self.size // 2))
-    
-    def get_rect(self):
-        return pygame.Rect(self.x - self.size // 2, self.y - self.size // 2, self.size, self.size)
 
 # Funkce pro kontrolu kolize hráče s medkitem
 def check_medkit_collision(player_x, player_y, player_radius):
